pyproject.toml
[tool.poetry]
name = "mlops-project"
version = "0.1.0"
description = "MLOps project for training and deploying a classification model."
authors = ["Yannick <yannick.vanneste@student.howest.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "^1.2.2"
pandas = "^2.0.1"
numpy = "^1.24.3"
imblearn = "^0.0"
joblib = "^1.2.0"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
flask = "^2.3.2"
python-dotenv = "^1.0.0"
azure-ai-ml = "^1.6.0"
azure-identity = "^1.13.0"
requests = "^2.31.0"
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
